Further restructuring and build & deploy setup
