--- atomic-agents/pyproject.toml
@@ -1,35 +1,0 @@
-[build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "atomic_agents"
-version = "0.3.7"
-description = "A versatile framework designed to facilitate the creation and management of intelligent agents."
-authors = [
-    { name = "Kenny Vaneetvelde", email = "kenny.vaneetvelde@gmail.com" }
-]
-readme = "README.md"
-license = { file = "LICENSE" }
-dependencies = [
-    "beautifulsoup4>=4.12.3,<5.0.0",
-    "google_api_python_client>=2.114.0,<3.0.0",
-    "instructor>=1.3.4,<2.0.0",
-    "markdownify>=0.12.1,<0.13.0",
-    "openai>=1.35.12,<2.0.0",
-    "pydantic>=2.8.0,<3.0.0",
-    "PyPDF2>=3.0.1,<4.0.0",
-    "Requests>=2.32.3,<3.0.0",
-    "rich>=13.7.1,<14.0.0",
-]
-
-[project.urls]
-homepage = "https://github.com/KennyVaneetvelde/atomic_agents"
-repository = "https://github.com/KennyVaneetvelde/atomic_agents"
-
-[tool.setuptools.packages.find]
-where = ["."]
-include = ["atomic_agents"]
-
-[tool.black]
-line-length = 127
