--- atomic-agents/requirements.txt
@@ -1,23 +1,0 @@
-# development dependencies
-black
-flake8
-pytest
-pytest-cov
-
-# library dependencies
-beautifulsoup4>=4.12.3,<5.0.0
-faiss_cpu>=1.8.0,<2.0.0
-google_api_python_client>=2.114.0,<3.0.0
-groq>=0.9.0,<0.10.0
-instructor>=1.3.4,<2.0.0
-markdownify>=0.12.1,<0.13.0
-numpy>=1.26.4,<2.0.0
-openai>=1.35.12,<2.0.0
-playwright>=1.45.0,<2.0.0
-pydantic>=2.8.0,<3.0.0
-PyPDF2>=3.0.1,<4.0.0
-pytest>=8.2.2,<9.0.0
-Requests>=2.32.3,<3.0.0
-rich>=13.7.1,<14.0.0
-sympy>=1.12,<2.0.0
-youtube_transcript_api>=0.6.2,<0.7.0
