--- atomic-assembler/atomic_assembler/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "1.0.0"
