+++ atomic-forge/atomic_tools/calculator/pyproject.toml
@@ -1,20 +1,22 @@
-[project]
+[tool.poetry]
 name = "calculator"
 version = "1.0"
+description = ""
+authors = ["Kenny Vaneetvelde <kenny.vaneetvelde@gmail.com>"]
 readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-    "atomic-agents",
-    "pydantic>=2.8.2,<3",
-    "sympy>=1.12,<2",
-]
+package-mode = false
 
-[tool.uv]
-dev-dependencies = [
-    "coverage>=7.6.1",
-    "pytest-cov>=5.0.0",
-    "pytest>=8.3.3",
-]
+[tool.poetry.dependencies]
+python = ">=3.12,<4"
+atomic-agents = {path = "../../../", develop = true}
+pydantic = ">=2.8.2,<3"
+sympy = ">=1.12,<2"
 
-[tool.uv.sources]
-atomic-agents = { path = "../../../atomic-agents" }
+[tool.poetry.group.dev.dependencies]
+coverage = ">=7.6.1,<8.0.0"
+pytest-cov = ">=5.0.0,<6.0.0"
+pytest = ">=8.3.3,<9.0.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
