--- atomic-forge/atomic_tools/calculator/requirements.txt
@@ -1,4 +1,0 @@
-atomic_agents>=0.3.3,<0.4
-pydantic>=2.8.2,<3
-sympy>=1.12,<2
-
