--- atomic-forge/atomic_tools/youtube_transcript_scraper/requirements.txt
@@ -1,4 +1,0 @@
-atomic_agents>=0.3.6,<0.4
-google_api_python_client>=2.114.0,<3
-pydantic>=2.9.1,<3
-youtube_transcript_api>=0.6.2,<1
