+++ pyproject.toml
@@ -1,46 +1,45 @@
 [build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
 
-[project]
+[tool.poetry]
 name = "atomic-agents"
-version = "1.0.0"
-description = "A versatile framework for creating and managing intelligent agents, with an optional CLI tool."
-authors = [
-    { name = "Kenny Vaneetvelde", email = "kenny.vaneetvelde@gmail.com" }
-]
+version = "1.0.5"
+description = "A versatile framework for creating and managing intelligent agents."
+authors = ["Kenny Vaneetvelde <kenny.vaneetvelde@gmail.com>"]
 readme = "README.md"
-license = { file = "LICENSE" }
-requires-python = ">=3.12"
-dependencies = [
-    "beautifulsoup4>=4.12.3,<5.0.0",
-    "google_api_python_client>=2.114.0,<3.0.0",
-    "instructor>=1.3.4,<2.0.0",
-    "markdownify>=0.12.1,<0.13.0",
-    "openai>=1.35.12,<2.0.0",
-    "pydantic>=2.8.0,<3.0.0",
-    "PyPDF2>=3.0.1,<4.0.0",
-    "Requests>=2.32.3,<3.0.0",
-    "rich>=13.7.1,<14.0.0",
+license = "LICENSE"
+packages = [
+    { include = "atomic_agents", from = "atomic-agents" },
+    { include = "atomic_assembler", from = "atomic-assembler" }
 ]
 
-[project.optional-dependencies]
-cli = [
-    "pyfiglet>=1.0.2,<2.0.0",
-    "textual>=0.77.0,<0.78.0",
-    "gitpython>=3.1.43,<3.2.0",
-    "pyyaml>=6.0.2,<7.0.0",
-]
+[tool.poetry.dependencies]
+python = ">=3.12,<4.0"
+instructor = ">=1.3.4,<2.0.0"
+openai = ">=1.35.12,<2.0.0"
+pydantic = ">=2.8.0,<3.0.0"
+rich = ">=13.7.1,<14.0.0"
+toml = ">=0.10.2,<1.0.0"
+gitpython = ">=3.1.43,<4.0.0"
+pyfiglet = ">=1.0.2,<2.0.0"
+textual = ">=0.82.0,<1.0.0"
+pyyaml = ">=6.0.2,<7.0.0"
 
-[project.scripts]
+[tool.poetry.group.dev.dependencies]
+black = ">=24.8.0,<25.0.0"
+build = ">=1.2.2,<2.0.0"
+flake8 = ">=7.1.1,<8.0.0"
+pytest = ">=8.3.3,<9.0.0"
+pytest-cov = ">=5.0.0,<6.0.0"
+fawltydeps = ">=0.17.0"
+
+[tool.poetry.scripts]
 atomic = "atomic_assembler.main:main"
 
-[project.urls]
+[tool.poetry.urls]
 homepage = "https://github.com/KennyVaneetvelde/atomic_agents"
 repository = "https://github.com/KennyVaneetvelde/atomic_agents"
-
-[tool.hatch.build.targets.wheel]
-packages = ["atomic-agents/atomic_agents", "atomic-assembler/atomic_assembler"]
 
 [tool.black]
 line-length = 127
