+++ requirements.txt
@@ -1,0 +1,5 @@
+instructor>=1.3.4,<2.0.0
+openai>=1.35.12,<2.0.0
+pydantic>=2.8.0,<3.0.0
+rich>=13.7.1,<14.0.0
+toml>=0.10.2,<1.0.0
